refactor BookList component
