+++ src/containers/BookList.js
@@ -1,11 +1,38 @@
+import React from 'react';
 import { connect } from 'react-redux';
+import PropTypes from 'prop-types';
 
-import Books from '../components/Books';
+import Book from '../components/Book';
+
+const BookList = ({ books }) => {
+  const bookItems = books.map(book => <Book key={book.id} book={book} />);
+
+  return (
+    <table className="table table-hover w-75 shadow-lg mb-4 rounded border">
+      <thead>
+        <tr>
+          <th scope="col">Book ID</th>
+          <th scope="col">Title</th>
+          <th scope="col">Category</th>
+        </tr>
+      </thead>
+      <tbody>{bookItems}</tbody>
+    </table>
+  );
+};
+
+BookList.propTypes = {
+  books: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.number,
+      title: PropTypes.string,
+      category: PropTypes.string,
+    }),
+  ).isRequired,
+};
 
 const mapStateToProps = state => ({
   books: state.books,
 });
 
-const BookList = connect(mapStateToProps, null)(Books);
-
-export default BookList;
+export default connect(mapStateToProps, null)(BookList);
