+++ src/reducers/books.js
@@ -1,8 +1,6 @@
 import { CREATE_BOOK, REMOVE_BOOK } from '../actions/types';
 
-const initialState = [];
-
-const bookReducer = (state = initialState, action) => {
+const bookReducer = (state = [], action) => {
   switch (action.type) {
     case CREATE_BOOK:
       return [...state, action.payload];
